Camera 1 Settings moved to separate activity
Splitting up settings
