+++ app/build.gradle
@@ -27,10 +27,8 @@ android {
 }
 
 dependencies {
-
     implementation "androidx.core:core:1.9.0"  //needed for broadcast receiver
     implementation 'androidx.appcompat:appcompat:1.5.1'
     implementation 'com.google.android.material:material:1.7.0'
     implementation 'androidx.preference:preference:1.2.1'
-
 }
