+++ app/src/main/java/at/andreasrohner/spartantimelapserec/MainActivity.java
@@ -41,6 +41,7 @@ import androidx.core.content.ContextCompat;
 
 import at.andreasrohner.spartantimelapserec.data.RecSettings;
 import at.andreasrohner.spartantimelapserec.sensor.MuteShutter;
+import at.andreasrohner.spartantimelapserec.settings.LegacyCamera1SettingsCommon;
 
 public class MainActivity extends AppCompatActivity implements ForegroundService.statusListener {
 
@@ -85,7 +86,7 @@ public class MainActivity extends AppCompatActivity implements ForegroundService
 		}
 		Context context = getApplicationContext();
 		SharedPreferences prefs = PreferenceManager.getDefaultSharedPreferences(context);
-		SettingsCommon.setDefaultValues(context, prefs);
+		LegacyCamera1SettingsCommon.setDefaultValues(context, prefs);
 		if (GithubStar.shouldShowStarDialog(this)) GithubStar.starDialog(this,"https://github.com/woheller69/timelapsecamera");
 	}
