+++ app/src/main/res/xml/camera1_preferences.xml
@@ -1,8 +1,7 @@
-<PreferenceScreen xmlns:app="http://schemas.android.com/apk/res-auto">
-    <PreferenceCategory
-        app:title="@string/pref_restapi_settings_header">
-
-<!--
+<PreferenceScreen xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:layout_width="fill_parent"
+    android:layout_height="fill_parent">
 
     <PreferenceCategory android:title="@string/pref_camera_settings_category">
         <at.andreasrohner.spartantimelapserec.preference.CameraInformation
@@ -112,10 +111,6 @@
             android:dialogMessage="@string/pref_schedule_recording"
             android:key="pref_schedule_recording"
             android:title="@string/pref_schedule_recording" />
-
-    </PreferenceCategory>
-
- -->
 
     </PreferenceCategory>
 </PreferenceScreen>
